bigclasses_backend/backend/enrollments/serializers.py
from rest_framework import serializers
class EnrollmentSerializer(serializers.Serializer):
    full_name = serializers.CharField(max_length=255)
    email = serializers.EmailField()
    phone = serializers.CharField(max_length=15)
